Title Read Gardasil data from csv file

Title Read Gardasil data from csv file

read_gardasil(other.as.NA = FALSE)

Arguments

other.as.NA
If FALSE (default), unknown/other race is treated as its own category. If TRUE, treat unknown/other race as NA.

Value

Cleaned/formatted data.frame

Details

The other.as.NA option is provided because the original manuscript was unclear about how other/unknown race was treated. Reproducing the reported results requires this parameter to be FALSE, but its coefficients are not reported.

Examples

gardasil2 <- read_gardasil()
summary(gardasil2)

#>       Age         AgeGroup              Race         Shots       Completed
#>  Min.   :11.00   18-26:712   white        :732   Min.   :1.000   no :944  
#>  1st Qu.:15.00   11-17:701   black        :443   1st Qu.:1.000   yes:469  
#>  Median :18.00               hispanic     : 52   Median :2.000            
#>  Mean   :18.55               other/unknown:186   Mean   :2.069            
#>  3rd Qu.:22.00                                   3rd Qu.:3.000            
#>  Max.   :26.00                                   Max.   :3.000            
#>             InsuranceType MedAssist                             Location  
#>  medical assistance:275   yes: 275   Odenton                        :798  
#>  private payer     :723   no :1138   White Marsh                    :165  
#>  hospital based    : 84              Johns Hopkins Outpatient Center: 89  
#>  military          :331              Bayview                        :361  
#>                                                                           
#>                                                                           
#>    LocationType          PracticeType   RaceSummary 
#>  urban   :450   pediatric      :515   white   :732  
#>  suburban:963   family practice:365   minority:495  
#>                 OB-GYN         :533   NA's    :186  
#>                                                     
#>                                                     
#>